requirements.txt
# Flask
flask>=3.0.0

# Database
flask-sqlalchemy>=3.1.0
psycopg2-binary>=2.9.9

# Environment and config
python-dotenv>=1.0.0

# HTTP (Brevo API)
requests>=2.28.0

# YouTube transcript (no API key; uses captions when available)
youtube-transcript-api>=0.6.2

# Production WSGI server (optional)
gunicorn>=21.0.0
